address feedback: re-organize usb codes
